// src/api/back/models/UpdateRuleIndexDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Public REST api for dota2classic
 * All stuff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateRuleIndexDto
 */
export interface UpdateRuleIndexDto {
    /**
     * 
     * @type {number}
     * @memberof UpdateRuleIndexDto
     */
    ruleId: number;
    /**
     * 
     * @type {number}
     * @memberof UpdateRuleIndexDto
     */
    parent?: number | null;
    /**
     * 
     * @type {number}
     * @memberof UpdateRuleIndexDto
     */
    index?: number;
}

export function UpdateRuleIndexDtoFromJSON(json: any): UpdateRuleIndexDto {
    return UpdateRuleIndexDtoFromJSONTyped(json, false);
}

export function UpdateRuleIndexDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateRuleIndexDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'ruleId': json['ruleId'],
        'parent': !exists(json, 'parent') ? undefined : json['parent'],
        'index': !exists(json, 'index') ? undefined : json['index'],
    };
}

export function UpdateRuleIndexDtoToJSON(value?: UpdateRuleIndexDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'ruleId': value.ruleId,
        'parent': value.parent,
        'index': value.index,
    };
}